update gzip task to check is file

+++ server/gzip.ts
@@ -1,4 +1,4 @@
-import { readFileSync, writeFileSync } from "fs";
+import { readFileSync, writeFileSync, lstatSync } from "fs";
 import { resolve } from "path";
 import { gzipSync } from "zlib";
 import * as glob from "glob";
@@ -6,7 +6,9 @@ import * as glob from "glob";
 const publicPath = resolve(__dirname, "../public");
 const gzippable = glob.sync(`${publicPath}/**/?(*.html|*.js|*.css)`);
 gzippable.forEach((file: any) => {
-  const content = readFileSync(file);
-  const zipped = gzipSync(content);
-  writeFileSync(`${file}`, zipped);
+  if(lstatSync(file).isFile()) {
+    const content = readFileSync(file);
+    const zipped = gzipSync(content);
+    writeFileSync(`${file}`, zipped);
+  }
 });
